// src/app/constants/enums/wp.enum.ts
// 兔肉 1
// 鸽肉 2
// 蛇胆
// 松鼠肉
// 熊掌
// 烤鸭
// 蟠桃
// 金桔
// 绸缎
// 麦穗
// 普洱
// 断肠草
// 黄精
// 续断
// 景天
// 灵芝
// 情花
// 龙胆
// 人参
// 无名精
// 鹤顶红
// 金疮药
// 行军丹
// 小还丹
// 大还丹
// 无极金丹
// 逍遥散
// 承气饮
// 七星海棠
// 悲酥清风
// 十全十美
// 九花玉露丸
export enum WpIdEnum {
  TuRou = 1,
  GeRou = 2,
  SheDan = 3,
  SongShuRou = 4,
  XiongZhang = 5,
  KaoYa = 6,
  PanTao = 7,
  JinJu = 8,
  ChouDuan = 9,
  MaiSui = 10,
  PuEr = 11,
  DuanChangCao = 12,
  HuangJing = 13,
  XuDuan = 14,
  JingTian = 15,
  LingZhi = 16,
  QingHua = 17,
  LongDan = 18,
  RenShen = 19,
  WuMingJing = 20,
  HeDingHong = 21,
  JinChuangYao = 22,
  XingJunDan = 23,
  XiaoHuanDan = 24,
  DaHuanDan = 25,
  WuJiJinDan = 26,
  XiaoYaoSan = 27,
  ChengQiYin = 28,
  QiXingHaiTang = 29,
  BeiSuQingFeng = 30,
  ShiQuanShiMei = 31,
  JiuHuaYuLuWan = 32,
}